feat: Dashboard User
