+++ src/app/(main)/dashboard/profile/_components/MyAccount.tsx
@@ -1,0 +1,315 @@
+"use client";
+import Button from "@/components/Button";
+import Input from "@/components/Input";
+import Seperator from "@/components/Seperator";
+import { Spinner } from "@/components/ui/Spinner";
+import {
+    GetUserProfile,
+    UpdateUserProfile,
+    uploadProfilePicture,
+} from "@/services/api/user";
+import { useSession } from "next-auth/react";
+import Image from "next/image";
+import React, { useState } from "react";
+import Swal from "sweetalert2";
+import User from "@/models/user";
+
+export default function MyAccount() {
+    const session = useSession();
+
+    const [loading, setLoading] = useState(true);
+    const [userData, setUserData] = useState<User>();
+    const [username, setUsername] = useState<string>("");
+    const [firstName, setFirstName] = useState<string>("");
+    const [middleName, setMiddleName] = useState<string>("");
+    const [lastName, setLastName] = useState<string>("");
+    const [email, setEmail] = useState<string>("");
+    const [major, setMajor] = useState<string>("");
+    const [batch, setBatch] = useState<string>("");
+    const [profilePicture, setProfilePicture] = useState<File | null>(null);
+    const [activeTab, setActiveTab] = useState("My Account");
+
+    // Fetch user data
+    React.useEffect(() => {
+        const fetchData = async () => {
+            if (session.data == null) {
+                return;
+            }
+            try {
+                const userData = await GetUserProfile(
+                    session.data.user.id,
+                    session.data.user.access_token
+                );
+                setUserData(userData);
+                setUsername(userData.username);
+                setFirstName(userData.first_name);
+                setMiddleName(userData.middle_name || "");
+                setLastName(userData.last_name);
+                setEmail(userData.email);
+                setMajor(userData.major);
+                setBatch(userData.year);
+                setLoading(false);
+            } catch (error) {
+                console.error("Error fetching user data:", error);
+            }
+        };
+
+        fetchData().then((r) => r);
+    }, [session.data]);
+
+    const handleProfilePictureChange = (
+        e: React.ChangeEvent<HTMLInputElement>
+    ) => {
+        const file = e.target.files?.[0];
+        if (file) {
+            setProfilePicture(file);
+        }
+    };
+
+    const handleProfilePictureUpload = async () => {
+        if (profilePicture) {
+            if (!session.data) {
+                return;
+            }
+            try {
+                const updatedUser = await uploadProfilePicture(
+                    profilePicture,
+                    session.data.user.access_token
+                );
+                setUserData(updatedUser);
+                setProfilePicture(null);
+
+                await Swal.fire({
+                    icon: "success",
+                    title: "Profile Picture Updated",
+                    showConfirmButton: false,
+                    timer: 1500,
+                });
+
+                window.location.reload();
+            } catch (error) {
+                await Swal.fire({
+                    icon: "error",
+                    title: "Error",
+                    text: "Error uploading profile picture",
+                });
+                console.error("Error uploading profile picture:", error);
+            }
+        }
+    };
+
+    const handleSave = async () => {
+        if (!session.data) {
+            return;
+        }
+        try {
+            const updatedUser = await UpdateUserProfile(
+                username,
+                firstName,
+                middleName,
+                lastName,
+                email,
+                major,
+                batch,
+                session.data.user.access_token
+            );
+
+            setUserData(updatedUser);
+
+            await Swal.fire({
+                icon: "success",
+                title: "Profile Updated",
+                showConfirmButton: false,
+                timer: 1500,
+            });
+
+            window.location.reload();
+        } catch (error) {
+            console.error("Error updating user profile:", error);
+        }
+    };
+
+    if (loading) {
+        return (
+            <div className="flex h-screen items-center justify-center">
+                <Spinner className="text-sky-500">
+                    <span className="text-sky-500">Loading...</span>
+                </Spinner>
+            </div>
+        );
+    }
+
+    return (
+        <section className="px-4">
+            <div className="grid grid-cols-1 gap-5 p-4 lg:grid-cols-2">
+                <div>
+                    <div className="rounded-lg border border-[#CBCBCB] bg-white shadow-lg">
+                        <div className="mt-2 px-6 py-3">
+                            <p className="text-[16px] font-[500]">
+                                Personal Information
+                            </p>
+                        </div>
+                        <Seperator className="border-gray-100" />
+                        <div className="mt-2 space-y-6 px-6 pb-6">
+                            {/*Username*/}
+                            <div>
+                                <Input
+                                    htmlFor="username"
+                                    label="Username"
+                                    type="text"
+                                    value={username}
+                                    placeholder={userData?.username}
+                                    onChange={(e) =>
+                                        setUsername(e.target.value)
+                                    }
+                                />
+                            </div>
+                            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
+                                <Input
+                                    htmlFor="first-name"
+                                    label="First Name"
+                                    type="text"
+                                    value={firstName}
+                                    placeholder={userData?.first_name}
+                                    onChange={(e) =>
+                                        setFirstName(e.target.value)
+                                    }
+                                />
+                                <Input
+                                    htmlFor="middle-name"
+                                    label="Middle Name*"
+                                    type="text"
+                                    value={middleName}
+                                    placeholder={userData?.middle_name}
+                                    onChange={(e) =>
+                                        setMiddleName(e.target.value)
+                                    }
+                                />
+                            </div>
+                            <div>
+                                <Input
+                                    htmlFor="last-name"
+                                    label="Last Name"
+                                    type="text"
+                                    value={lastName}
+                                    placeholder={userData?.last_name}
+                                    onChange={(e) =>
+                                        setLastName(e.target.value)
+                                    }
+                                />
+                            </div>
+                            <div>
+                                <Input
+                                    htmlFor="email-address"
+                                    label="Email Address"
+                                    type="email"
+                                    value={email}
+                                    placeholder={userData?.email}
+                                    onChange={(e) => setEmail(e.target.value)}
+                                />
+                            </div>
+                            <div className="mt-16 flex justify-end space-x-2 py-5">
+                                <Button
+                                    onClick={handleSave}
+                                    className="border-[#02ABF3] bg-[#02ABF3] px-8 py-2 text-white hover:bg-white hover:text-[#02ABF3]"
+                                >
+                                    Save
+                                </Button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div>
+                    <div
+                        className="rounded-lg border border-[#CBCBCB] bg-white shadow-lg"
+                        style={{ height: "fit-content" }}
+                    >
+                        <div className="mt-2 px-6 py-3">
+                            <p className="text-[16px] font-[500]">
+                                Profile Photo
+                            </p>
+                        </div>
+                        <Seperator className="border-gray-100" />
+                        <div className="p-6">
+                            <div className="mb-4">
+                                <div className="flex gap-3">
+                                    <div className="">
+                                        <Image
+                                            src={
+                                                userData?.profile_picture ||
+                                                "https://sg.pufacomputing.live/Assets/male.jpeg"
+                                            }
+                                            alt={`${userData?.first_name} ${userData?.last_name} Profile Picture's`}
+                                            className="h-16 w-16 rounded-full object-cover"
+                                            width={480}
+                                            height={240}
+                                        />
+                                    </div>
+                                    <div className="flex flex-col items-start gap-1">
+                                        <button
+                                            onClick={() => {
+                                                document
+                                                    .getElementById(
+                                                        "dropzone-file"
+                                                    )
+                                                    ?.click();
+                                            }}
+                                            className="font-medium"
+                                        >
+                                            Edit your photo
+                                        </button>
+                                    </div>
+                                </div>
+                            </div>
+                            <div className="text-[#353535]">
+                                <label
+                                    htmlFor="dropzone-file"
+                                    className="mx-auto mt-2 flex w-full max-w-xl cursor-pointer flex-col items-center rounded-xl border-2 border-dashed border-gray-300 p-6 hover:border-gray-400 dark:border-gray-700"
+                                >
+                                    <svg
+                                        xmlns="http://www.w3.org/2000/svg"
+                                        fill="none"
+                                        viewBox="0 0 24 24"
+                                        strokeWidth="1.5"
+                                        stroke="currentColor"
+                                        className="h-8 w-8 text-gray-500 dark:text-gray-400"
+                                    >
+                                        <path
+                                            strokeLinecap="round"
+                                            strokeLinejoin="round"
+                                            d="M12 16.5V9.75m0 0l3 3m-3-3l-3 3M6.75 19.5a4.5 4.5 0 01-1.41-8.775 5.25 5.25 0 0110.233-2.33 3 3 0 013.758 3.848A3.752 3.752 0 0118 19.5H6.75z"
+                                        />
+                                    </svg>
+
+                                    <h2 className="mt-1 font-medium tracking-wide">
+                                        Add Profile Picture
+                                    </h2>
+
+                                    <p className="mt-2 text-xs tracking-wide ">
+                                        Upload or darg & drop your file SVG,
+                                        PNG, or JPG.{" "}
+                                    </p>
+
+                                    <input
+                                        id="dropzone-file"
+                                        type="file"
+                                        className="hidden"
+                                        onChange={handleProfilePictureChange}
+                                    />
+                                </label>
+                            </div>
+                            <div className="mt-6 flex justify-end space-x-2">
+                                <Button
+                                    onClick={handleProfilePictureUpload}
+                                    className="border-[#02ABF3] bg-[#02ABF3] px-8 py-2 text-white hover:bg-white hover:text-[#02ABF3]"
+                                >
+                                    Save
+                                </Button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </section>
+    );
+}
